requirements.txt
rerun-sdk>=0.15.0
numpy>=1.20.0
pillow>=8.0.0
opencv-python>=4.5.0
scipy>=1.7.0
torch>=1.9.0
tqdm>=4.60.0

